Support DDL deparse of the rest commands.
    CREATE/DROP ACCESS METHOD
    CREATE/ALTER/DROP AGGREGATE
    CREATE/ALTER/DROP CAST
    CREATE/ALTER/DROP COLLATION
    COMMENT
    CREATE/ALTER/DROP CONVERSION
    ALTER DEFAULT PRIVILEGES
    CREATE/ALTER/DROP DOMAIN
    CREATE/ALTER/DROP EXTENSION
    CREATE/ALTER/DROP FOREIGN DATA WRAPPER
    CREATE/ALTER/DROP FOREIGN TABLE
    CREATE/ALTER FUNCTION/PROCEDURE
    GRANT
    CREATE/DROP INDEX
    CREATE/ALTER/DROP LANGUAGE
    CREATE/ALTER/DROP MATERIALIZED VIEW
    CREATE/DROP OPERATOR [CLASS/FAMILY]
    CREATE/ALTER/DROP PUBLICATION
    CREATE/ALTER/DROP POLICY
    REFRESH MATERIALIZED VIEW
    REVOKE
    ALTER/DROP ROUTINE
    CREATE/ALTER/DROP RULE
    CREATE/ALTER/DROP SCHEMA
    SECURITY LABEL
    CREATE/ALTER/DROP SEQUENCE
    CREATE/ALTER/DROP SERVER
    CREATE/ALTER/DROP STATISTICS
    CREATE/ALTER/DROP TEXT SEARCH CONFIGURATION/PARSER/DICTIONARY/TEMPLATE
    CREATE/DROP TRANSFORM
    CREATE/ALTER TRIGGER
    CREATE/ALTER/DROP TYPE
    CREATE/ALTER/DROP USER MAPPING
    CREATE/ALTER/DROP VIEW -- Note #4

(Note #3) Removed an undesirable 'on' from the identity string for TRANSFORM
in getObjectIdentityParts. This is needed to make deparse of DROP
TRANSFORM command work since 'on' is not present in the current
DROP TRANSFORM syntax. For example, the correct syntax is
DROP TRANSFORM trf FOR INT LANGUAGE SQL;
instead of DROP TRANSFORM trf FOR INT ON LANGUAGE SQL;

(Note #4) For VIEWSTMT, we created pg_get_viewdef_internal to get the
SELECT query of the pending VIEWSTMT. In the case that the CREATE VIEW command
is still in progress, we would need to search the system cache RULERELNAME to get the
rewrite rule of the view as oppose to querying pg_rewrite as in pg_get_viewdef_worker(),
the latter will return empty result.